Move leaderboard table into its own component
